+++ src/docs/devguide/hugo/site/content/release-notes.md
@@ -2,7 +2,7 @@ This page describes changes to the Money Transfer Retailer Interface implemented
 
 ## v2.11.0
 
-Released xx May 2019
+Released 21 May 2019
 
 - Added the following new operations:
     - Orders Resources:
@@ -14,7 +14,7 @@ Released xx May 2019
     - `MoneyTransferOrderCancelResponse` - describes a response for a cancel order request.
     - `MoneyTransferCancelOrderReversal` - describes a request and response for a cancel order reversal.
 
-- Fixed relative path constants by removing an extra `/` introduced in `v2.6.0` for the following:
+- Fixed relative path constants by removing an extra `/` introduced in `v2.6.0` for the following operations:
     - `confirmPayment`
     - `reversePayment`
     - `orderStatus`