+++ src/docs/devguide/hugo/site/content/release-notes.md
@@ -1,5 +1,23 @@
 This page describes changes to the Money Transfer Retailer Interface implemented across different releases of the interface.
 
+## v2.11.0
+
+Released 21 May 2019
+
+- Added the following new operations:
+    - Orders Resources:
+        - CancelOrder: used to cancel an order.
+        - CancelOrderReversal: used to reverse a cancel order request.
+
+- Added the following new models:
+    - `MoneyTransferOrderCancelRequest` - describes a request to cancel an order.
+    - `MoneyTransferOrderCancelResponse` - describes a response for a cancel order request.
+    - `MoneyTransferCancelOrderReversal` - describes a request and response for a cancel order reversal.
+
+- Fixed relative path constants by removing an extra `/` introduced in `v2.6.0` for the following operations:
+    - `confirmPayment`
+    - `reversePayment`
+    - `orderStatus`
 
 ## v2.11.0